refactor(canister-core): merge certification + certifier + nested_tree into cert/
Three flat modules — `certification` (stateless primitives + the response tree),
`certifier` (the stateful engine), and `nested_tree` (a generic Merkle tree) —
were confusingly named and leaked their internals crate-wide. Fold them into one
`cert/` module with private submodules:

- cert/nested_tree.rs — the generic tree. Used only by primitives, so now fully
  private to cert (was a crate-wide pub module).
- cert/primitives.rs — the IC-certification vocabulary (paths, hashes, cert
  expressions) + the CertifiedResponses tree wrapper.
- cert/certifier.rs — the Certifier engine.

cert/mod.rs re-exports only the narrow interface the rest of the crate needs
(the addressing/hashing helpers used by asset/redirect/protection, plus
Certifier). The tree wrapper (CertifiedResponses), its addressing internals
(RequestHash, WitnessResult), and the whole nested_tree are now private to cert
— they no longer appear in any other module's imports.

Resolves the certification/certifier name collision (there's now one `cert`) and
shrinks the crate's cross-module surface. Pure code motion; submodule renames are
git-tracked.

Verified: builds (incl. canbench-rs), 158 unit tests pass, fmt + clippy clean,
full workspace builds, e2e 26/26 pass.
