+++ crates/canister-core/src/http.rs
@@ -1,4 +1,4 @@
-use crate::certification::{
+use crate::cert::{
     build_ic_certificate_expression_from_headers, build_ic_certificate_expression_header,
 };
 use candid::{CandidType, Deserialize};
